+++ playwright/fixtures/pages/fronts.ts
@@ -43,7 +43,7 @@ const tagPages: TagPage[] = [
 	{
 		path: getTestUrl({
 			stage,
-			path: '/world/americas',
+			path: '/tone/recipes/all',
 			type: 'tagPage',
 			adtest: 'fixed-puppies',
 		}),
@@ -60,4 +60,14 @@ const frontWithPageSkin: Front = {
 	section: 'uk',
 };
 
-export { frontWithPageSkin, fronts, tagPages };
+const frontWithExclusion: Front = {
+	path: getTestUrl({
+		stage,
+		path: '/us-news/baltimore-bridge-collapse',
+		type: 'tagPage',
+		adtest: 'clear',
+	}),
+	section: 'us-news',
+};
+
+export { frontWithPageSkin, fronts, tagPages, frontWithExclusion };