+++ src/init/consented/comscore.spec.ts
@@ -1,16 +1,9 @@
-import {
-	getConsentFor,
-	onConsent,
-} from '@guardian/consent-management-platform';
-import type { ConsentState } from '@guardian/consent-management-platform/dist/types';
-import type { TCFv2ConsentState } from '@guardian/consent-management-platform/dist/types/tcfv2';
-import { loadScript } from '@guardian/libs';
+import type { ConsentState, TCFv2ConsentState } from '@guardian/libs';
+import { getConsentFor, loadScript, onConsent } from '@guardian/libs';
 import { commercialFeatures } from 'lib/commercial-features';
 import { _ } from './comscore';
 
 const { setupComscore } = _;
-
-jest.mock('@guardian/consent-management-platform');
 
 const SOURCEPOINT_ID = '5efefe25b8e05c06542b2a77';
 
@@ -80,6 +73,8 @@ const AusWithConsent = {
 jest.mock('@guardian/libs', () => ({
 	loadScript: jest.fn(() => Promise.resolve()),
 	log: jest.fn(),
+	onConsent: jest.fn(),
+	getConsentFor: jest.fn(),
 }));
 
 jest.mock('lib/commercial-features', () => ({