+++ src/insert/comments-expanded-advert.ts
@@ -66,6 +66,7 @@ const getRightColumn = (): HTMLElement => {
 	const selector = window.guardian.config.isDotcomRendering
 		? '.commentsRightColumn'
 		: '.js-discussion__ad-slot';
+
 	const rightColumn: HTMLElement | null = document.querySelector(selector);
 
 	if (!rightColumn) throw new Error('Could not find right column.');
@@ -179,13 +180,20 @@ export const initCommentsExpandedAdverts = (): Promise<void> => {
 		const currentBreakpoint = getBreakpoint(getViewport().width);
 		if (currentBreakpoint === 'mobile') {
 			if (
-				window.guardian.config.isDotcomRendering &&
-				!window.guardian.config.switches.mobileDiscussionAds
+				window.guardian.config.tests?.commercialMegaTestControl ===
+				'control'
 			) {
 				return;
 			}
 			void handleCommentsLoadedMobileEvent();
 		} else {
+			if (
+				!window.guardian.config.isDotcomRendering &&
+				window.guardian.config.tests?.commercialMegaTestControl ===
+					'control'
+			) {
+				return;
+			}
 			void handleCommentsLoadedEvent();
 		}
 	});
@@ -195,6 +203,12 @@ export const initCommentsExpandedAdverts = (): Promise<void> => {
 	 * we need to remove the existing slots and create new slots.
 	 */
 	document.addEventListener('comments-state-change', () => {
+		if (
+			window.guardian.config.tests?.commercialMegaTestControl ===
+			'control'
+		) {
+			return;
+		}
 		void removeMobileCommentsExpandedAds();
 	});