+++ src/insert/spacefinder/article.ts
@@ -6,7 +6,6 @@ import {
 	createAdSlot,
 	wrapSlotInContainer,
 } from 'core/create-ad-slot';
-import { isInSectionAdDensityVariant } from 'experiments/utils';
 import { spaceFiller } from 'insert/spacefinder/space-filler';
 import type {
 	RuleSpacing,
@@ -20,10 +19,9 @@ import {
 } from 'lib/detect/detect-breakpoint';
 import { waitForAdvert } from '../../lib/dfp/wait-for-advert';
 import fastdom from '../../utils/fastdom-promise';
-import { mediator } from '../../utils/mediator';
-import { fillDynamicAdSlot } from '../fill-dynamic-advert-slot';
 import { computeStickyHeights, insertHeightStyles } from '../sticky-inlines';
 import { initCarrot } from './carrot-traffic-driver';
+import { isInHighValueSection } from './utils';
 
 type SlotName = Parameters<typeof createAdSlot>[0];
 
@@ -36,10 +34,13 @@ const hasImages = !!window.guardian.config.page.lightboxImages?.images.length;
 const hasShowcaseMainElement =
 	window.guardian.config.page.hasShowcaseMainElement;
 
-const isInAdDensityVariant = isInSectionAdDensityVariant();
+const isInMegaTestControl =
+	window.guardian.config.tests?.commercialMegaTestControl === 'control';
+
+const increaseAdDensity = isInHighValueSection && !isInMegaTestControl;
 
 const minDistanceBetweenRightRailAds = 500;
-const minDistanceBetweenInlineAds = isInAdDensityVariant ? 500 : 750;
+const minDistanceBetweenInlineAds = increaseAdDensity ? 500 : 750;
 
 /**
  * Rules to avoid inserting ads in the right rail too close to each other
@@ -74,14 +75,13 @@ const inlineAdSlotContainerRules: Record<string, RuleSpacing> = {
 const getStickyContainerClassname = (i: number) =>
 	`${adSlotContainerClass}-${i + 2}`;
 
-const insertAdAtPara = (
+const insertSlotAtPara = async (
 	para: Node,
 	name: string,
 	type: SlotName,
 	classes?: string,
-	sizes?: SizeMapping,
 	containerOptions: ContainerOptions = {},
-): Promise<void> => {
+): Promise<HTMLElement> => {
 	const ad = createAdSlot(type, {
 		name,
 		classes,
@@ -89,17 +89,21 @@ const insertAdAtPara = (
 
 	const node = wrapSlotInContainer(ad, containerOptions);
 
-	return fastdom
-		.mutate(() => {
-			if (para.parentNode) {
-				para.parentNode.insertBefore(node, para);
-			}
-		})
-		.then(async () => {
-			const shouldForceDisplay = ['im', 'carrot'].includes(name);
-			await fillDynamicAdSlot(ad, shouldForceDisplay, sizes);
-		});
+	await fastdom.mutate(() => {
+		if (para.parentNode) {
+			para.parentNode.insertBefore(node, para);
+		}
+	});
+
+	return ad;
 };
+
+type FillAdSlot = (
+	name: string,
+	slot: HTMLElement,
+	/* additional sizes to be added to the slot (used only by googletag) */
+	additionalSizes?: SizeMapping,
+) => Promise<void>;
 
 /**
  * Decide whether we have enough space to add additional sizes for a given advert.
@@ -136,11 +140,11 @@ const decideAdditionalSizes = async (
 	);
 };
 
-const addDesktopInline1 = (): Promise<boolean> => {
+const addDesktopInline1 = (fillSlot: FillAdSlot): Promise<boolean> => {
 	const tweakpoint = getCurrentTweakpoint();
 	const hasLeftCol = ['leftCol', 'wide'].includes(tweakpoint);
 
-	let ignoreList = ` > :not(p):not(h2):not(ul):not(.${adSlotContainerClass}):not(#sign-in-gate):not(.sfdebug)`;
+	let ignoreList = `:scope > :not(p):not(h2):not(ul):not(.${adSlotContainerClass}):not(#sign-in-gate):not(.sfdebug)`;
 	if (hasLeftCol) {
 		ignoreList +=
 			':not([data-spacefinder-role="richLink"]):not([data-spacefinder-role="thumbnail"])';
@@ -150,16 +154,16 @@ const addDesktopInline1 = (): Promise<boolean> => {
 
 	const rules: SpacefinderRules = {
 		bodySelector: articleBodySelector,
-		candidateSelector: ' > p',
+		candidateSelector: ':scope > p, [data-spacefinder-role="nested"] > p',
 		minAbove: isImmersive ? 700 : 300,
 		minBelow: 300,
 		opponentSelectorRules: {
 			// don't place ads right after a heading
-			' > h2': {
-				minAboveSlot: isInAdDensityVariant ? 150 : 5,
-				minBelowSlot: isInAdDensityVariant ? 0 : 190,
+			':scope > h2, [data-spacefinder-role="nested"] > h2': {
+				minAboveSlot: increaseAdDensity ? 150 : 5,
+				minBelowSlot: increaseAdDensity ? 0 : 190,
 			},
-			[` .${adSlotContainerClass}`]: {
+			[`.${adSlotContainerClass}`]: {
 				minAboveSlot: 500,
 				minBelowSlot: 500,
 			},
@@ -167,11 +171,11 @@ const addDesktopInline1 = (): Promise<boolean> => {
 				minAboveSlot: 35,
 				minBelowSlot: 400,
 			},
-			' [data-spacefinder-role="immersive"]': {
+			'[data-spacefinder-role="immersive"]': {
 				minAboveSlot: 0,
 				minBelowSlot: 600,
 			},
-			' figure.element--supporting': {
+			'figure.element--supporting': {
 				minAboveSlot: 500,
 				minBelowSlot: 0,
 			},
@@ -187,13 +191,9 @@ const addDesktopInline1 = (): Promise<boolean> => {
 
 	const insertAd: SpacefinderWriter = async (paras) => {
 		const slots = paras.slice(0, 1).map(async (para) => {
-			return insertAdAtPara(
-				para,
-				'inline1',
-				'inline',
-				'inline',
-				additionalSizes,
-			);
+			const name = 'inline1';
+			const slot = await insertSlotAtPara(para, name, 'inline', 'inline');
+			await fillSlot(name, slot, additionalSizes);
 		});
 
 		await Promise.all(slots);
@@ -209,7 +209,7 @@ const addDesktopInline1 = (): Promise<boolean> => {
 /**
  * Inserts all inline ads on desktop except for inline1.
  */
-const addDesktopRightRailAds = (): Promise<boolean> => {
+const addDesktopRightRailAds = (fillSlot: FillAdSlot): Promise<boolean> => {
 	let minAbove = 1000;
 
 	/**
@@ -229,12 +229,12 @@ const addDesktopRightRailAds = (): Promise<boolean> => {
 	const largestSizeForSlot = adSizes.halfPage.height;
 	const rules: SpacefinderRules = {
 		bodySelector: articleBodySelector,
-		candidateSelector: ' > p',
+		candidateSelector: ':scope > p, [data-spacefinder-role="nested"] > p',
 		minAbove,
 		minBelow: 300,
 		opponentSelectorRules: {
 			...rightRailAdSlotContainerRules,
-			' [data-spacefinder-role="immersive"]': {
+			'[data-spacefinder-role="immersive"]': {
 				minAboveSlot: 0,
 				minBelowSlot: 600,
 			},
@@ -289,14 +289,15 @@ const addDesktopRightRailAds = (): Promise<boolean> => {
 				),
 			};
 
-			return insertAdAtPara(
+			const slot = await insertSlotAtPara(
 				para,
 				`inline${i + 2}`,
 				'inline',
 				'inline',
-				additionalSizes,
 				containerOptions,
 			);
+
+			return fillSlot(`inline${i + 2}`, slot, additionalSizes);
 		});
 
 		await Promise.all(slots);
@@ -309,33 +310,56 @@ const addDesktopRightRailAds = (): Promise<boolean> => {
 	});
 };
 
-const addMobileInlineAds = (): Promise<boolean> => {
-	const minDistanceFromArticleTop = isInAdDensityVariant ? 100 : 200;
+const addMobileInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
+	const minDistanceFromArticleTop = !isInMegaTestControl ? 100 : 200;
 
-	const rules: SpacefinderRules = {
+	/**
+	 * These 2 sets of rules are for the changes to "ranked" articles as part of the mega test
+	 */
+	const oldRules: SpacefinderRules = {
 		bodySelector: articleBodySelector,
-		candidateSelector: [
-			' > p',
-			' > h2',
-			' > [data-spacefinder-type$="NumberedTitleBlockElement"]',
-		],
+		candidateSelector: ':scope > p',
+		minAbove: 200,
+		minBelow: 200,
+		opponentSelectorRules: {
+			':scope > h2': {
+				minAboveSlot: 100,
+				minBelowSlot: 250,
+			},
+			...inlineAdSlotContainerRules,
+			[`:scope > :not(p):not(h2):not(.${adSlotContainerClass}):not(#sign-in-gate)`]:
+				{
+					minAboveSlot: 35,
+					minBelowSlot: 200,
+				},
+		},
+		filter: (candidate, lastWinner) => {
+			if (!lastWinner) {
+				return true;
+			}
+			const distanceBetweenAds = candidate.top - lastWinner.top;
+			return distanceBetweenAds >= minDistanceBetweenInlineAds;
+		},
+	};
+
+	const ignoreList = `:not(p):not(h2):not(hr):not(.${adSlotContainerClass}):not(#sign-in-gate):not([data-spacefinder-type$="NumberedTitleBlockElement"])`;
+
+	const newRules: SpacefinderRules = {
+		bodySelector: articleBodySelector,
+		candidateSelector:
+			':scope > p, :scope > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"], [data-spacefinder-role="nested"] > p',
 		minAbove: minDistanceFromArticleTop,
 		minBelow: 200,
 		opponentSelectorRules: {
 			// don't place ads right after a heading
-			' > h2': {
-				minAboveSlot: 100,
-				minBelowSlot: 0,
-			},
-			// these are just fancy headings
-			' > [data-spacefinder-type$="NumberedTitleBlockElement"]': {
-				minAboveSlot: 100,
-				minBelowSlot: 0,
-			},
-
+			':scope > h2, [data-spacefinder-role="nested"] > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"]':
+				{
+					minAboveSlot: 100,
+					minBelowSlot: 0,
+				},
 			...inlineAdSlotContainerRules,
 			// this is a catch-all for elements that are not covered by the above rules, these will generally be things like videos, embeds and atoms. minBelowSlot is higher to push ads a bit further down after these elements
-			[` > :not(p):not(h2):not(hr):not(.${adSlotContainerClass}):not(#sign-in-gate):not([data-spacefinder-type$="NumberedTitleBlockElement"])`]:
+			[`:scope > ${ignoreList}, [data-spacefinder-role="nested"] > ${ignoreList}`]:
 				{
 					minAboveSlot: 35,
 					minBelowSlot: 200,
@@ -358,21 +382,24 @@ const addMobileInlineAds = (): Promise<boolean> => {
 		},
 	};
 
+	const rules = isInMegaTestControl ? oldRules : newRules;
+
 	const insertAds: SpacefinderWriter = async (paras) => {
-		const slots = paras.map((para, i) =>
-			insertAdAtPara(
-				para,
-				i === 0 ? 'top-above-nav' : `inline${i}`,
-				i === 0 ? 'top-above-nav' : 'inline',
-				'inline',
+		const slots = paras.map(async (para, i) => {
+			const name = i === 0 ? 'top-above-nav' : `inline${i}`;
+			const type = i === 0 ? 'top-above-nav' : 'inline';
+			const slot = await insertSlotAtPara(para, name, type, 'inline');
+			return fillSlot(
+				name,
+				slot,
 				// Add the mobile portrait interstitial size to inline1 and inline2
 				i == 1 || i == 2
 					? {
 							mobile: [adSizes.portraitInterstitial],
 					  }
 					: undefined,
-			),
-		);
+			);
+		});
 		await Promise.all(slots);
 	};
 
@@ -383,45 +410,58 @@ const addMobileInlineAds = (): Promise<boolean> => {
 	});
 };
 
-const addInlineAds = (): Promise<boolean> => {
+/**
+ * Add inline slots to the article body
+ * @param fillSlot A function to call that will fill the slot when each ad slot has been inserted,
+ * these could be google display ads or opt opt consentless ads.
+ */
+const addInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
 	const isMobile = getCurrentBreakpoint() === 'mobile';
 	if (isMobile) {
-		return addMobileInlineAds();
+		return addMobileInlineAds(fillSlot);
 	}
 
 	if (isPaidContent) {
-		return addDesktopRightRailAds();
+		return addDesktopRightRailAds(fillSlot);
 	}
 
 	// Add the rest of the inline ad slots after a position for inline1 has been found.
 	// We don't wan't inline1 and inline2 targeting the same paragraph.
-	return addDesktopInline1().then(() => addDesktopRightRailAds());
+	return addDesktopInline1(fillSlot).then(() =>
+		addDesktopRightRailAds(fillSlot),
+	);
 };
 
-const attemptToAddInlineMerchAd = (): Promise<boolean> => {
+const attemptToAddInlineMerchAd = (
+	fillAdSlot: FillAdSlot,
+): Promise<boolean> => {
 	const breakpoint = getCurrentBreakpoint();
 	const isMobileOrTablet = breakpoint === 'mobile' || breakpoint === 'tablet';
 
 	const rules: SpacefinderRules = {
 		bodySelector: articleBodySelector,
-		candidateSelector: ' > p',
+		candidateSelector: ':scope > p',
 		minAbove: 300,
 		minBelow: 300,
 		opponentSelectorRules: {
-			' > .merch': {
+			':scope > .merch': {
 				minAboveSlot: 0,
 				minBelowSlot: 0,
 			},
-			' > header': {
+			':scope > header': {
 				minAboveSlot: isMobileOrTablet ? 300 : 700,
 				minBelowSlot: 0,
 			},
-			' > h2': {
+			':scope > h2': {
 				minAboveSlot: 100,
 				minBelowSlot: 250,
 			},
+			':scope > #sign-in-gate': {
+				minAboveSlot: 0,
+				minBelowSlot: 400,
+			},
 			...inlineAdSlotContainerRules,
-			[` > :not(p):not(h2):not(.${adSlotContainerClass}):not(#sign-in-gate):not(.sfdebug)`]:
+			[`:scope > :not(p):not(h2):not(.${adSlotContainerClass}):not(#sign-in-gate):not(.sfdebug)`]:
 				{
 					minAboveSlot: 200,
 					minBelowSlot: 400,
@@ -429,22 +469,17 @@ const attemptToAddInlineMerchAd = (): Promise<boolean> => {
 		},
 	};
 
-	const insertAds: SpacefinderWriter = (paras) => {
+	const insertAds: SpacefinderWriter = async (paras) => {
 		if (typeof paras[0] === 'undefined') {
 			throw new Error(
 				'Trying to insert inline merch before a node that does not exist',
 			);
 		}
-		return insertAdAtPara(
-			paras[0],
-			'im',
-			'im',
-			'',
-			{},
-			{
-				className: 'ad-slot-container--im',
-			},
-		);
+		const slot = await insertSlotAtPara(paras[0], 'im', 'im', '', {
+			className: 'ad-slot-container--im',
+		});
+
+		await fillAdSlot('im', slot);
 	};
 
 	return spaceFiller.fillSpace(rules, insertAds, {
@@ -454,34 +489,26 @@ const attemptToAddInlineMerchAd = (): Promise<boolean> => {
 	});
 };
 
-const doInit = async (): Promise<boolean> => {
+/**
+ * Init all the article body adverts, including `im` and `carrot`
+ * @param fillAdSlot a function to fill the ad slots
+ */
+const init = async (fillAdSlot: FillAdSlot): Promise<boolean> => {
 	if (!commercialFeatures.articleBodyAdverts) {
 		return Promise.resolve(false);
 	}
 
+	await addInlineAds(fillAdSlot);
+
 	const im = window.guardian.config.page.hasInlineMerchandise
-		? attemptToAddInlineMerchAd()
+		? attemptToAddInlineMerchAd(fillAdSlot)
 		: Promise.resolve(false);
 	const inlineMerchAdded = await im;
 	if (inlineMerchAdded) await waitForAdvert('dfp-ad--im');
-
-	await addInlineAds();
 
 	await initCarrot();
 
 	return im;
 };
 
-/**
- * Initialise article body ad slots
- */
-export const init = (): Promise<boolean> => {
-	// Also init when the main article is redisplayed
-	// For instance by the signin gate.
-	mediator.on('page:article:redisplayed', doInit);
-	// DCR doesn't have mediator, so listen for CustomEvent
-	document.addEventListener('article:sign-in-gate-dismissed', () => {
-		void doInit();
-	});
-	return doInit();
-};
+export { init, addInlineAds, type FillAdSlot };