+++ src/types/global.ts
@@ -1,5 +1,5 @@
-import type { VendorName } from '@guardian/consent-management-platform/dist/types';
 import type { CustomClaims } from '@guardian/identity-auth';
+import type { VendorName } from '@guardian/libs';
 import type { EventTimer } from '../core/event-timer';
 import type { PageTargeting } from '../core/targeting/build-page-targeting';
 import type {
@@ -259,6 +259,9 @@ type ComscoreGlobals = {
 	c2: string;
 	cs_ucfr: string;
 	comscorekw?: string;
+	options?: {
+		enableFirstPartyCookie?: boolean;
+	};
 };
 
 type CustomIdTokenClaims = CustomClaims & {